Move ControlInfo and ValueInfo to components
