+++ frontend/src/modules/dashboard/components/ControlInfo.tsx
