+++ frontend/src/modules/dashboard/components/ValueInfo.tsx
